增加了bean validation框架

+++ beanvalidation/src/main/java/com/jfireframework/beanvalidation/constraint/Max.java
@@ -17,5 +17,5 @@ import java.lang.annotation.Target;
  */
 public @interface Max
 {
-    public int value();
+    public long value();
 }
