+++ beanvalidation/src/main/java/com/jfireframework/beanvalidation/constraint/Min.java
@@ -17,5 +17,5 @@ import java.lang.annotation.Target;
  */
 public @interface Min
 {
-    public int value();
+    public long value();
 }
